Improved DRYness of the 'record tests' handling across test runners
Better reuse promotes consistency
