+++ launchable/test_runners/go_test.py
@@ -14,9 +14,4 @@ def subset(client):
     client.run()
 
 
-@click.argument('source_roots', required=True, nargs=-1)
-@launchable.record.tests
-def record_tests(client, source_roots):
-    for root in source_roots:
-        client.scan(root, "*.xml")
-    client.run()
+record_tests = launchable.CommonRecordTestImpls(__name__).report_files()
