+++ launchable/test_runners/minitest.py
@@ -32,13 +32,4 @@ def subset(client, files):
     client.run()
 
 
-@click.argument('report_dirs', required=True, nargs=-1)
-@launchable.record.tests
-def record_tests(client, report_dirs):
-    for root in report_dirs:
-        if os.path.isdir(root):
-            client.scan(root, "*.xml")
-        else:
-            client.report(root)
-
-    client.run()
+record_tests = launchable.CommonRecordTestImpls(__name__).report_files()
